added TOTP support and more logging

+++ src/userapi/user_routers.py
@@ -9,6 +9,8 @@ from django.core.cache import cache
 from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 
+from .enums.account_status import AccountStatus
+
 from .enums.log_severity import LogSeverity
 
 from .models import Account, ServiceLog, SessionToken, EmailHistoryLog
@@ -38,7 +40,6 @@ def request_user_info(request):
         return JsonResponse({'status': 'ERROR'}, status=404)
 
     data = json.loads(request.body)
-
     session_token = request.headers.get('session-token', '')
     user_agent = request.META['HTTP_USER_AGENT']
     username = data.get('username', '')
@@ -46,10 +47,14 @@ def request_user_info(request):
     account = validate_session_token(username, user_agent, session_token)
     if account is None:
         return JsonResponse({'status': 'BAD_SESSION_TOKEN'}, status=401)
-
+    
+    if not account.authenticated:
+        return JsonResponse({'status': 'NOT_AUTHENTICATED'}, status=401)
+    if account.status != AccountStatus.OK:
+        return JsonResponse({'status': 'ACCOUNT_DISABLED'}, status=401)
+    
     email = request_decrypt(account.id, account.encrypted_email, account.id)
     user_data = {
-        'status': 'SUCCESS',
         'username': account.username,
         'email': email,
         'totpEnabled': account.totp_enabled,
@@ -85,6 +90,12 @@ def request_user_info(request):
             'isSelfSession': is_self_session
         })
 
+    log = ServiceLog.objects.create(
+        content=f"{account.username} ({account.id}) requested user information!",
+        severity=LogSeverity.VERBOSE
+    )
+    log.save()
+
     return JsonResponse({
         'status': "SUCCESS",
         'userData': user_data,
@@ -108,6 +119,10 @@ def change_email(request):
     account = validate_session_token(username, user_agent, session_token)
     if account is None:
         return JsonResponse({'status': 'BAD_SESSION_TOKEN'}, status=401)
+    if not account.authenticated:
+        return JsonResponse({'status': 'NOT_AUTHENTICATED'}, status=401)
+    if account.status != AccountStatus.OK:
+        return JsonResponse({'status': 'ACCOUNT_DISABLED'}, status=401)
     
     has_request_change_recently = cache.get(f"email_verification_{str(account.id)}")
     if has_request_change_recently is not None:
@@ -115,17 +130,17 @@ def change_email(request):
 
     # Check TOTP if required
     if account.totp_enabled and not totp_code:
-        return JsonResponse({'status': 'MISSING_TOTP'}, status=401)
+        return JsonResponse({'status': 'MISSING_TOTP'}, status=403)
     if account.totp_enabled:
         verify_result = verify_totp_code(account.id, totp_code)
         verify_status = verify_result['status']
         if verify_status != 'SUCCESS':
             return JsonResponse({'status': 'ERROR'}, status=500)
         if not verify_result['valid']:
-            return JsonResponse({'status': 'BAD_TOTP'}, status=401)
+            return JsonResponse({'status': 'BAD_TOTP'}, status=403)
         
     if not verify_password(account.hashed_password, password):
-        return JsonResponse({'status': 'BAD_CURRENT_PASSWORD'}, status=401)
+        return JsonResponse({'status': 'BAD_CURRENT_PASSWORD'}, status=403)
     
     if not is_valid_email(new_email):
         return JsonResponse({'status': 'INVALID_EMAIL'}, status=400)
@@ -150,6 +165,12 @@ def change_email(request):
         if not response['success']:
             return JsonResponse({'status': 'ERROR'}, status=500)
         cache.set(f"email_verification_{str(account.id)}", True, timeout=360)
+
+        log = ServiceLog.objects.create(
+            content=f"{account.username} ({account.id}) has to change email",
+            severity=LogSeverity.VERBOSE
+        )
+        log.save()
     except Exception as e:
         log_message = f"Tried to send new change confirmation email for {account.username} ({account.id}), but failed due to :: {e}\n\n{traceback.format_exc()}"
         print(log_message, flush=True)
@@ -181,13 +202,18 @@ def confirm_email_change(request):
     if account is None:
         return JsonResponse({'status': 'BAD_SESSION_TOKEN'}, status=401)
     
+    if not account.authenticated:
+        return JsonResponse({'status': 'NOT_AUTHENTICATED'}, status=401)
+    if account.status != AccountStatus.OK:
+        return JsonResponse({'status': 'ACCOUNT_DISABLED'}, status=401)
+    
     try:
         result = verify_email_verification_token(tokenID, token)
         if not result['valid']:
-            return JsonResponse({'status': 'INVALID_TOKEN'}, status=401)
+            return JsonResponse({'status': 'INVALID_TOKEN'}, status=403)
         
         if not str(account.id) == result['accountID']:
-            return JsonResponse({'status': 'INVALID_TOKEN'}, status=401)
+            return JsonResponse({'status': 'INVALID_TOKEN'}, status=403)
     except Exception as e:
         print(e, flush=True)
         log = ServiceLog.objects.create(
@@ -221,6 +247,12 @@ def confirm_email_change(request):
     try:
         email = request_decrypt(account.id, old_encrypted_email, account.id)
         send_email_with_content(email, 'Email Changed', 'This is a notice that you have changed your email account for Muimi! If this is not intended, please reply to contact admin!')
+    
+        log = ServiceLog.objects.create(
+            content=f"{account.username} ({account.id}) has confirmed email change!",
+            severity=LogSeverity.LOG
+        )
+        log.save()
     except Exception as e:
         print(e, flush=True)
         log = ServiceLog.objects.create(
@@ -251,20 +283,25 @@ def change_password(request):
     if account is None:
         return JsonResponse({'status': 'BAD_SESSION_TOKEN'}, status=401)
     
+    if not account.authenticated:
+        return JsonResponse({'status': 'NOT_AUTHENTICATED'}, status=401)
+    if account.status != AccountStatus.OK:
+        return JsonResponse({'status': 'ACCOUNT_DISABLED'}, status=401)
+
     # Check TOTP if required
     if account.totp_enabled and not totp_code:
-        return JsonResponse({'status': 'MISSING_TOTP'}, status=401)
+        return JsonResponse({'status': 'MISSING_TOTP'}, status=403)
     if account.totp_enabled:
         verify_result = verify_totp_code(account.id, totp_code)
         verify_status = verify_result['status']
         if verify_status != 'SUCCESS':
             return JsonResponse({'status': 'ERROR'}, status=500)
         if not verify_result['valid']:
-            return JsonResponse({'status': 'BAD_TOTP'}, status=401)
+            return JsonResponse({'status': 'BAD_TOTP'}, status=403)
     
     # Check passwording
     if not verify_password(account.hashed_password, current_password):
-        return JsonResponse({'status': 'BAD_CURRENT_PASSWORD'}, status=401)
+        return JsonResponse({'status': 'BAD_CURRENT_PASSWORD'}, status=403)
     
     if not is_valid_password(password):
         return JsonResponse({'status': 'BAD_PASSWORD'}, status=400)
@@ -305,8 +342,20 @@ def revoke_session(request):
     account = validate_session_token(username, user_agent, session_token)
     if account is None:
         return JsonResponse({'status': 'BAD_SESSION_TOKEN'}, status=401)
-
+    
+    if not account.authenticated:
+        return JsonResponse({'status': 'NOT_AUTHENTICATED'}, status=401)
+    if account.status != AccountStatus.OK:
+        return JsonResponse({'status': 'ACCOUNT_DISABLED'}, status=401)
+    
     SessionToken.objects.filter(account=account, id=session_id).delete()
+
+    log = ServiceLog.objects.create(
+        content=f"{account.username} ({account.id}) removed a session token!",
+        severity=LogSeverity.LOG
+    )
+    log.save()
+
     return JsonResponse({
         'status': 'SUCCESS'
     })
